tests/wpt/css-tests/cssom-view-1_dev/xhtml1/matchMediaAddListener.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
        <title>CSS Test: CSSOM View matchMedia addListener</title>
        <link href="mailto:pwx.frontend@gmail.com" rel="author" title="Chris Wu" />
        <link href="http://www.w3.org/TR/cssom-view/#the-mediaquerylist-interface" rel="help" />
        <meta content="dom" name="flags" />
        <script src="/resources/testharness.js" type="text/javascript">
        &lt;script src="/resources/testharnessreport.js" type="text/javascript" /&gt;
        &lt;style type="text/css"&gt;
            iframe { border: none; }
        &lt;/style&gt;
    &lt;/head&gt;
    &lt;body&gt;
        &lt;div id="log"&gt;&lt;/div&gt;
        &lt;iframe width="200" height="100" id="iframe1" &gt;&lt;/iframe&gt;
        &lt;script&gt;
            function reflow(doc) {
                doc.body.offsetWidth;
            }

            var iframe = document.querySelector("iframe");
            var iframe_window = window.frames[0];
            var iframe1 = document.getElementById("iframe1");
            reflow(iframe_window.document);

            var i = 0;
            var totalCount = 10;
            var count = 0;
            var divineCount = 10;
            var width_list = [201,199];
            var mq1 = iframe_window.matchMedia("(max-width:200px)");
            mq1.addListener(function(mql){

                count = count + 1;

            });

            var equalAssert = async_test("Check for the correct number of event triggers");

            var changeFrameWidth = function(iWidth) {
                iframe1.style.width = iWidth + "px";
                i = (i === 0) ? 1 : 0;
                totalCount = totalCount - 1;
                if(totalCount &gt; 0)
                {
                    setTimeout(function(){
                        changeFrameWidth(width_list[i]);
                    }, 100);
                }
                else
                {
                    setTimeout(function(){
                        equalAssert.step(function(){
                            assert_equals(divineCount, count, "this will be 10 times of event triggers by change width");
                        });
                        equalAssert.done();
                    }, 100);
                }
            };

            changeFrameWidth(width_list[0]);


        </script>
    </head><body>

</body></html>
